Move testes AdicionarFluxoView para arquivo próprio #63
